int back to byte in structs
git-svn-id: https://trac.beirdo.ca:4443/svn/havokmud/trunk@552 69fa473d-82f3-0310-a870-f1ebe01cfa2b

+++ havokmud/structs.h
@@ -1306,7 +1306,7 @@ struct weaponskills
 #endif
 struct char_special_data
 {
-  int spellfail;        /* max # for spell failure (101) */
+  byte spellfail;        /* max # for spell failure (101) */
   byte tick;             /* the tick that the mob/player is on  */
   byte pmask;            /* poof mask                           */
   byte position;         /* Standing or ...                        */
